experiments/requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.0.4
prompt-toolkit==3.0.27
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.14.1
pycryptodome-test-vectors==1.0.7
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pyzmq==22.3.0
requests==2.27.1
setuptools==60.9.3
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
gunicorn==20.1.0